+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 Luminati Proxy manager - Change Log
 
+## 1.109.867 Stable
+- :boom: All proxies will listen on 127.0.0.1 by default, unless specific interface is set or ips whitelist is defined
+- :sparkles: Add IP address from which log in was performed to whitelist ip list
+
 ## 1.109.493 Stable
 - :bug: Fixed socks proxy security issue